Fix Ghidra symbol postprocessing
